requirements.txt
django-ninja==1.4.5
django-cors-headers==4.9.0